infiniband-diags/ibdiag_common.c: Eliminate compile warning on x86_64 archs
src/ibdiag_common.c: In function pretty_print
src/ibdiag_common.c:95: warning: field precision should have type int,
but argument 3 has type long int

+++ src/ibdiag_common.c
@@ -92,7 +92,7 @@ static void pretty_print(int start, int width, const char *str)
 		}
 		if (e - str == 1)
 			e = p;
-		fprintf(stderr, "%.*s\n%*s", e - str, str, start, "");
+		fprintf(stderr, "%.*s\n%*s", (int)(e - str), str, start, "");
 		str = e;
 	}
 }
